Avancée du state

+++ src/shared/state/Property.cpp
@@ -9,4 +9,4 @@ bool Property::operator==(const Property &other) const{
 }
 
 
-Property::Property(int position, std::string name, int valuePurchase) : position(position), name(name), valuePurchase(valuePurchase){}
+Property::Property(int position, std::string name, long long valuePurchase) : position(position), name(name), valuePurchase(valuePurchase){}
